Improve helper method to set permalink structure
