+++ tests/codeception/_support/Helper/Acceptance.php
@@ -6,5 +6,5 @@ namespace Helper;
 
 class Acceptance extends \Codeception\Module
 {
-
+    use PermalinkTrait;
 }
