+++ tests/codeception/_support/Steps/Theme.php
@@ -21,15 +21,11 @@ trait Theme
      */
     public function iSeeTheLinkForAuthorInTheByline($authorSlug)
     {
+        $this->setPermalinkStructure('/%postname%/');
+
         $author = Author::get_by_term_slug($authorSlug);
 
-        if ($author->is_guest()) {
-            $authorId = $author->term_id * -1;
-        } else {
-            $authorId = $author->user_id;
-        }
-
-        $expectedAuthorLink = $_ENV['TEST_SITE_WP_URL'] . '/?author=' . $authorId;
+        $expectedAuthorLink = $_ENV['TEST_SITE_WP_URL'] . '/author/' . $author->slug . '/';
         $this->seeElement('.posted-by .byline a', ['href' => $expectedAuthorLink]);
     }
 }
