+++ tests/codeception/_support/WpunitTester.php
@@ -19,16 +19,4 @@
 class WpunitTester extends \Codeception\Actor
 {
     use _generated\WpunitTesterActions;
-
-   /**
-    * Define custom actions here
-    */
-   public function setPermalinkStructure($structure)
-   {
-       global $wp_rewrite;
-
-       $wp_rewrite->init();
-       $wp_rewrite->set_permalink_structure($structure);
-       $wp_rewrite->flush_rules(true);
-   }
 }
