+++ tests/codeception/_support/_generated/AcceptanceTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] 22922234d7424d54ed88212650bcca18
+<?php  //[STAMP] 78936fdd5db82d0fb05f2cd9cc26bd92
 namespace _generated;
 
 // This class was automatically generated by build task
@@ -8164,6 +8164,17 @@ trait AcceptanceTesterActions
      */
     public function amEditingPostWithId($id) {
         return $this->getScenario()->runStep(new \Codeception\Step\Condition('amEditingPostWithId', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Define custom actions here
+     * @see \Helper\Acceptance::setPermalinkStructure()
+     */
+    public function setPermalinkStructure($structure) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('setPermalinkStructure', func_get_args()));
     }
