+++ tests/codeception/_support/_generated/RestapiTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] 7a1c1071a1ae5d6ec28533a7c375c04b
+<?php  //[STAMP] 3eaae48a4d3216c424d19be8e4a8ed88
 namespace _generated;
 
 // This class was automatically generated by build task
@@ -5988,6 +5988,17 @@ trait RestapiTesterActions
      */
     public function startFollowingRedirects() {
         return $this->getScenario()->runStep(new \Codeception\Step\Action('startFollowingRedirects', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Define custom actions here
+     * @see \Helper\Restapi::setPermalinkStructure()
+     */
+    public function setPermalinkStructure($structure) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('setPermalinkStructure', func_get_args()));
     }
