+++ tests/codeception/_support/_generated/WpunitTesterActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] f433b0d5698bbe9f9463221977b46fe5
+<?php  //[STAMP] e9c7d4d655be75536dc3aa548fafe799
 namespace _generated;
 
 // This class was automatically generated by build task
@@ -404,6 +404,17 @@ trait WpunitTesterActions
      */
     public function echoLastQuery() {
         return $this->getScenario()->runStep(new \Codeception\Step\Action('echoLastQuery', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Define custom actions here
+     * @see \Helper\Wpunit::setPermalinkStructure()
+     */
+    public function setPermalinkStructure($structure) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('setPermalinkStructure', func_get_args()));
     }
